Transactions and msgpack serialization
* Add msgpack-c submodule

* Algorand transactions model
Follows the design of the SDKs

* Implementation and serialization of transactions
- All utilities for serialization to msgpack are located under src/algorand/transactions/util
- Also remove redundant constexpr const for global variables in inc/algorand/AlgorandConstants.hpp

* Add tests for msgpack serialization
Unpacking transaction is not possible by design, so the packing of transactions cannot be tested directly. Instead, the utility functions used to pack the transactions are tested with fake classes.

* Move constants to TransactionAdaptor
These constants will only be used for the encoding of transactions to msgpack, so they do not need to be visible for the entire project.